api/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
flask-cors==4.0.1
gunicorn==22.0.0
pandas==2.2.2
Faker==25.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
alembic==1.13.2